2021-11-29/23.42 MatrixPresentation
Первая редакция форматированного вывода матрицы в консоль.

+++ src/main/java/matrices/SquareMatrix.java
@@ -8,7 +8,7 @@ public class SquareMatrix extends Matrix {
         super(lines, lines);
     }
 
-    protected int findDeterminant(int[][] arr) {
+    protected long findDeterminant(int[][] arr) {
         int determinant = 0;
         if (arr.length == 1) {
             return arr[0][0];
